requirements.txt
aiobotocore==2.13.1
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
albucore==0.0.12
albumentations==1.4.11
annotated-types==0.7.0
anthropic==0.23.1
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-search-documents==11.4.0
azure-storage-blob==12.20.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bm25s==0.1.10
boto3==1.34.147
botocore==1.34.147
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
chromedriver-autoinstaller==0.6.4
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.1
cryptography==42.0.8
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.7
datasets==2.20.0
decorator==5.1.1
Deprecated==1.2.14
dill==0.3.8
dirtyjson==1.0.8
distro==1.9.0
dnspython==2.6.1
email_validator==2.2.0
eval_type_backport==0.2.0
executing==2.0.1
faiss-cpu==1.8.0.post1
fastapi==0.111.0
fastapi-cli==0.0.4
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
gkeepapi==0.15.1
google-api-core==2.19.1
google-api-python-client==2.137.0
google-auth==2.32.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
gpsoauth==1.1.1
greenlet==3.0.3
grpcio==1.65.0
grpcio-tools==1.65.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hyperframe==6.0.1
idna==3.7
imageio==2.34.2
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.26.0
isodate==0.6.1
jedi==0.19.1
jieba3k==0.35.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
kubernetes==30.1.0
lazy_loader==0.4
Levenshtein==0.25.1
lightning==2.3.3
lightning-utilities==0.11.5
llama-cloud==0.0.6
llama-index==0.10.54.post1
llama-index-agent-openai==0.2.8
llama-index-cli==0.1.12
llama-index-core==0.10.54
llama-index-embeddings-azure-openai==0.1.10
llama-index-embeddings-huggingface==0.2.2
llama-index-embeddings-openai==0.1.10
llama-index-indices-managed-llama-cloud==0.2.4
llama-index-legacy==0.9.48
llama-index-llms-azure-openai==0.1.8
llama-index-llms-openai==0.1.25
llama-index-multi-modal-llms-anthropic==0.1.4
llama-index-multi-modal-llms-azure-openai==0.1.4
llama-index-multi-modal-llms-openai==0.1.7
llama-index-program-openai==0.1.6
llama-index-question-gen-openai==0.1.3
llama-index-readers-azstorage-blob==0.1.7
llama-index-readers-file==0.1.30
llama-index-readers-google==0.2.9
llama-index-readers-llama-parse==0.1.6
llama-index-readers-nougat-ocr==0.1.3
llama-index-readers-s3==0.1.10
llama-index-readers-web==0.1.23
llama-index-retrievers-bm25==0.2.0
llama-index-storage-docstore-mongodb==0.1.3
llama-index-storage-docstore-redis==0.1.2
llama-index-storage-index-store-mongodb==0.1.2
llama-index-storage-kvstore-mongodb==0.1.2
llama-index-storage-kvstore-redis==0.1.5
llama-index-vector-stores-azureaisearch==0.1.9
llama-index-vector-stores-chroma==0.1.10
llama-index-vector-stores-faiss==0.1.2
llama-index-vector-stores-pinecone==0.1.7
llama-index-vector-stores-qdrant==0.2.11
llama-index-vector-stores-redis==0.2.0
llama-parse==0.4.6
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib==3.9.1
matplotlib-inline==0.1.7
mdurl==0.1.2
minijinja==2.0.1
mmh3==4.1.0
monotonic==1.6
motor==3.5.1
mpmath==1.3.0
msal==1.29.0
msal-extensions==1.2.0
multidict==6.0.5
multiprocess==0.70.16
munch==4.0.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
newspaper3k==0.2.8
nltk==3.8.1
nougat-ocr==0.1.17
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.18.1
openai==1.35.13
opencv-python-headless==4.10.0.84
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
orjson==3.10.6
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.4.0
pinecone-client==3.2.2
pinecone-plugin-interface==0.0.7
playwright==1.45.1
portalocker==2.10.0
posthog==3.5.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodomex==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
PyDrive==1.3.1
pyee==11.1.0
Pygments==2.18.0
PyJWT==2.8.0
pymongo==4.8.0
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySocks==1.7.1
PyStemmer==2.2.0.1
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-multipart==0.0.9
pytorch-lightning==2.3.3
pytz==2024.1
PyYAML==6.0.1
qdrant-client==1.10.1
rapidfuzz==3.9.4
redis==5.0.7
redisvl==0.1.3
regex==2024.5.15
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==2024.6.1
s3transfer==0.10.2
safetensors==0.4.3
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.0
sconf==0.2.5
selenium==4.23.1
sentence-transformers==3.0.1
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spider-client==0.0.27
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
striprtf==0.0.26
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tifffile==2024.7.2
tiktoken==0.7.0
timm==0.5.4
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.19.1
tomli==2.0.1
torch==2.3.1
torchmetrics==1.4.0.post0
torchvision==0.18.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
trio==0.26.0
trio-websocket==0.11.1
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.19
uvicorn==0.30.1
uvloop==0.19.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
